Revert https://github.com/punkave/sanitize-html/pull/226
The 'browser' field should be a node module intended for use in
a web browser, eg. using the `window` object or similar as per
https://docs.npmjs.com/files/package.json#browser . It's what
webpack imports if the key exists. This points it to browserified
source code which isn't what webpack wants.

Also the sanitize-html.js published to npm seems to be an empty
module, I don't know why this is yet, but either way I don't
believe this key should be here.