Update index.js
In case the input is not a string, would it be okay to just return the input?

I have a case, where its not certain what goes into the function. If a number goes in, or a number array, it will convert the numbers to strings which an cause problems for my project down the line.